refactor(swift): remove orphaned resource wrapper
