+++ apps/shared/OpenClawKit/Tests/OpenClawKitTests/ToolDisplayRegistryTests.swift
@@ -3,11 +3,6 @@ import Foundation
 import Testing
 
 @Suite struct ToolDisplayRegistryTests {
-    @Test func loadsToolDisplayConfigFromBundle() {
-        let url = OpenClawKitResources.bundle.url(forResource: "tool-display", withExtension: "json")
-        #expect(url != nil)
-    }
-
     @Test func resolvesKnownToolFromConfig() {
         let summary = ToolDisplayRegistry.resolve(name: "exec", args: nil)
         #expect(summary.emoji == "🛠️")
